New directory structure for conjure_core
