+++ conjure_oxide/src/lib.rs
@@ -9,8 +9,11 @@ pub mod rules;
 pub mod utils;
 
 pub use conjure_core::ast; // re-export core::ast as conjure_oxide::ast
-pub use conjure_core::ast::Model; // rexport core::ast::Model as conjure_oxide::Model
+pub use conjure_core::model::Model as Model; // rexport core::ast::Model as conjure_oxide::Model
 pub use conjure_core::solvers::Solver;
+pub use conjure_core::metadata;
+pub use conjure_core::rule;
+pub use conjure_rules::rule_set;
 pub use rules::eval_constant;
 
 pub use error::Error;
