+++ conjure_oxide/src/parse.rs
@@ -3,8 +3,9 @@ use std::collections::HashMap;
 use conjure_core::metadata::Metadata;
 use serde_json::Value;
 
-use crate::ast::{Constant, DecisionVariable, Domain, Expression, Model, Name, Range};
+use crate::ast::{Constant, DecisionVariable, Domain, Expression, Name, Range};
 use crate::error::{Error, Result};
+use crate::Model;
 use serde_json::Value as JsonValue;
 
 pub fn model_from_json(str: &str) -> Result<Model> {
