+++ conjure_oxide/src/rule_engine/rewrite.rs
@@ -1,9 +1,10 @@
 use crate::rule_engine::resolve_rules::{
-    get_rule_priorities, get_rules_vec, ResolveRulesError as ResolveError,
+    get_rule_priorities, get_rules_vec, ResolveRulesError as ResolveError
 };
-use conjure_core::ast::{Expression, Model};
-use conjure_core::rule::{Reduction, Rule};
-use conjure_rules::rule_set::RuleSet;
+use crate::ast::{Expression};
+use crate::rule::{Reduction, Rule};
+use crate::rule_set::RuleSet;
+use crate::Model;
 use std::fmt::Display;
 use thiserror::Error;
 use uniplate::uniplate::Uniplate;
