+++ conjure_oxide/src/solver/adaptors/sat_common.rs
@@ -3,7 +3,7 @@
 
 // (nd60, march 24) - i basically copied all this from @gskorokod's SAT implemention for the old
 // solver interface.
-use crate::{ast as conjure_ast, solver::SolverError, solver::SolverError::*};
+use crate::{ast as conjure_ast, solver::SolverError, solver::SolverError::*, Model as ConjureModel};
 use conjure_core::metadata::Metadata;
 use std::collections::HashMap;
 use thiserror::Error;
@@ -30,7 +30,7 @@ impl CNFModel {
         }
     }
 
-    pub fn from_conjure(conjure_model: conjure_ast::Model) -> Result<CNFModel, SolverError> {
+    pub fn from_conjure(conjure_model: ConjureModel) -> Result<CNFModel, SolverError> {
         let mut ans: CNFModel = CNFModel::new();
 
         for var in conjure_model.variables.keys() {
@@ -326,16 +326,15 @@ mod tests {
     use super::CNFModel;
     use crate::ast::Domain::{BoolDomain, IntDomain};
     use crate::ast::Expression::{And, Not, Or, Reference};
-    use crate::ast::{DecisionVariable, Model};
-    use crate::ast::{Expression, Name};
-    use crate::solver::SolverError;
+    use crate::ast::{Expression, Name, DecisionVariable};
+    use crate::{ast as conjure_ast, solver::SolverError, solver::SolverError::*, Model as ConjureModel};
     use crate::utils::testing::assert_eq_any_order;
 
     #[test]
     fn test_single_var() {
         // x -> [[1]]
 
-        let mut model: Model = Model::default();
+        let mut model: ConjureModel = ConjureModel::default();
 
         let x: Name = Name::UserName(String::from('x'));
         model.add_variable(x.clone(), DecisionVariable { domain: BoolDomain });
@@ -357,7 +356,7 @@ mod tests {
     fn test_single_not() {
         // Not(x) -> [[-1]]
 
-        let mut model: Model = Model::default();
+        let mut model: ConjureModel = ConjureModel::default();
 
         let x: Name = Name::UserName(String::from('x'));
         model.add_variable(x.clone(), DecisionVariable { domain: BoolDomain });
@@ -389,7 +388,7 @@ mod tests {
     fn test_single_or() {
         // Or(x, y) -> [[1, 2]]
 
-        let mut model: Model = Model::default();
+        let mut model: ConjureModel = ConjureModel::default();
 
         let x: Name = Name::UserName(String::from('x'));
         let y: Name = Name::UserName(String::from('y'));
@@ -430,7 +429,7 @@ mod tests {
     fn test_or_not() {
         // Or(x, Not(y)) -> [[1, -2]]
 
-        let mut model: Model = Model::default();
+        let mut model: ConjureModel = ConjureModel::default();
 
         let x: Name = Name::UserName(String::from('x'));
         let y: Name = Name::UserName(String::from('y'));
@@ -477,7 +476,7 @@ mod tests {
     fn test_multiple() {
         // [x, y] - equivalent to And(x, y) -> [[1], [2]]
 
-        let mut model: Model = Model::default();
+        let mut model: ConjureModel = ConjureModel::default();
 
         let x: Name = Name::UserName(String::from('x'));
         let y: Name = Name::UserName(String::from('y'));
@@ -510,7 +509,7 @@ mod tests {
     fn test_and() {
         // And(x, y) -> [[1], [2]]
 
-        let mut model: Model = Model::default();
+        let mut model: ConjureModel = ConjureModel::default();
 
         let x: Name = Name::UserName(String::from('x'));
         let y: Name = Name::UserName(String::from('y'));
@@ -548,7 +547,7 @@ mod tests {
     fn test_nested_ors() {
         // Or(x, Or(y, z)) -> [[1, 2, 3]]
 
-        let mut model: Model = Model::default();
+        let mut model: ConjureModel = ConjureModel::default();
 
         let x: Name = Name::UserName(String::from('x'));
         let y: Name = Name::UserName(String::from('y'));
@@ -599,7 +598,7 @@ mod tests {
     fn test_int() {
         // y is an IntDomain - only booleans should be allowed
 
-        let mut model: Model = Model::default();
+        let mut model: ConjureModel = ConjureModel::default();
 
         let x: Name = Name::UserName(String::from('x'));
         let y: Name = Name::UserName(String::from('y'));
@@ -623,7 +622,7 @@ mod tests {
     fn test_eq() {
         // Eq(x, y) - this operation is not allowed
 
-        let mut model: Model = Model::default();
+        let mut model: ConjureModel = ConjureModel::default();
 
         let x: Name = Name::UserName(String::from('x'));
         let y: Name = Name::UserName(String::from('y'));
