+++ conjure_oxide/src/utils/conjure.rs
@@ -1,7 +1,8 @@
 use crate::parse::model_from_json;
 use crate::utils::json::sort_json_object;
 use crate::Error as ParseErr;
-use conjure_core::ast::{Constant, Model, Name};
+use crate::ast::{Constant, Name};
+use crate::Model;
 use serde_json::{Map, Value as JsonValue};
 use std::collections::HashMap;
 use std::sync::{Arc, Mutex};
