+++ conjure_oxide/src/utils/testing.rs
@@ -2,8 +2,9 @@ use crate::utils::conjure::minion_solutions_to_json;
 use crate::utils::json::sort_json_object;
 use crate::utils::misc::to_set;
 use crate::Error;
-use conjure_core::ast::Name::UserName;
-use conjure_core::ast::{Constant, Model as ConjureModel, Name};
+use crate::ast::Name::UserName;
+use crate::ast::{Constant, Name};
+use crate::Model as ConjureModel;
 use serde_json::{Error as JsonError, Value as JsonValue};
 use std::collections::{HashMap, HashSet};
 use std::fmt::Debug;
