+++ conjure_oxide/tests/rewrite_tests.rs
@@ -11,6 +11,7 @@ use conjure_oxide::eval_constant;
 use conjure_oxide::rule_engine::{resolve_rules::resolve_rule_sets, rewrite::rewrite_model};
 use conjure_oxide::solver::{adaptors, Solver};
 use conjure_rules::{get_rule_by_name, get_rules};
+use conjure_oxide::Model;
 
 #[test]
 fn rules_present() {
