+++ crates/conjure_core/src/lib.rs
@@ -2,5 +2,6 @@ pub mod ast;
 pub mod metadata;
 pub mod rule;
 pub mod solvers;
+pub mod model;
 
 pub use solvers::Solver;
