+++ crates/conjure_core/src/rule.rs
@@ -1,10 +1,9 @@
 use std::fmt::{self, Display, Formatter};
 use std::hash::Hash;
-
 use thiserror::Error;
-
-use crate::ast::{Expression, Model, SymbolTable};
+use crate::ast::{Expression, SymbolTable};
 use crate::metadata::Metadata;
+use crate::model::Model;
 
 #[derive(Debug, Error)]
 pub enum ApplicationError {
